requirements.txt
gspread==6.1.2
protobuf==5.28.0
python-dotenv==1.0.1
Requests==2.32.3
